refactor: move stats to separate module
